CipherTalkV2/requirements.txt
pynacl
cryptography
sqlalchemy
typer
python-dotenv
pycryptodome